Move to an opentracing-mock artifact; add a README
