+++ opentracing-mock/src/main/java/io/opentracing/mock/MockTracer.java
@@ -1,4 +1,4 @@
-package io.opentracing.io.opentracing.mocktracer;
+package io.opentracing.mock;
 
 import io.opentracing.References;
 import io.opentracing.Span;
